Rename DBContext file name + Reload Migration
